+++ app.js
@@ -3,7 +3,7 @@ const logger = require('morgan')
 const cors = require('cors')
 const dotenv = require('dotenv')
 
-const contactsRouter = require('./routes/api/contacts')
+const contactsRouter = require('./routes/api/route')
 
 dotenv.config()
